# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user/add": {
      "post": {
        "operationId": "UserController_create",
        "summary": "create one new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/page": {
      "get": {
        "operationId": "UserController_findAll",
        "summary": "find all user by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasRole",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "summary": "find one user by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasRole",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/update": {
      "patch": {
        "operationId": "UserController_update",
        "summary": "update a user with given id and user object",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "user id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/disable{id}": {
      "delete": {
        "operationId": "UserController_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/enable/{id}": {
      "patch": {
        "operationId": "UserController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/role/add": {
      "post": {
        "operationId": "RoleController_create",
        "summary": "create one new role",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      }
    },
    "/role": {
      "get": {
        "operationId": "RoleController_findAll",
        "summary": "find role list by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasAuth",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      }
    },
    "/role/{id}": {
      "get": {
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      },
      "patch": {
        "operationId": "RoleController_update",
        "summary": "update role info",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      }
    },
    "/role/remove{id}": {
      "delete": {
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      }
    },
    "/role/enable/{id}": {
      "patch": {
        "operationId": "RoleController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      }
    },
...
    "/page": {
      "post": {
        "operationId": "PageController_create",
        "summary": "create one new page",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/page": {
      "get": {
        "operationId": "PageController_findAll",
        "summary": "find page list by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasAuth",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
    "/page/{id}": {
      "get": {
        "operationId": "PageController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      },
      "patch": {
        "operationId": "PageController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
...
    "/page/enable/{id}": {
      "patch": {
        "operationId": "PageController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Page"
        ]
      }
    },
...
    "/tag": {
      "post": {
        "operationId": "TagController_create",
        "summary": "create one new tag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tag"
        ]
      }
    },
    "/tag/page": {
      "get": {
        "operationId": "TagController_findAll",
        "summary": "find all tag by page",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasPost",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tag"
        ]
      }
    },
    "/tag/{id}": {
      "get": {
        "operationId": "TagController_findOne",
        "summary": "find one tag by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isDel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "hasPost",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tag"
        ]
      },
      "patch": {
        "operationId": "TagController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tag"
        ]
      }
    },
    "/tag/remove/{id}": {
      "delete": {
        "operationId": "TagController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tag"
        ]
      }
    },
    "/tag/enable/{id}": {
      "patch": {
        "operationId": "TagController_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tag"
        ]
      }
    },
...
  },
  "info": {
    "title": "My API",
    "description": "The description of my API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "scheme": "basic",
        "name": "access-token",
        "description": "user token authentication",
        "in": "headers"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "userAccount": {
            "type": "string",
            "description": "The userAccount must start with a uppercase letter and contain 5 to 12 letters and digits and underscores"
          },
          "userPassword": {
            "type": "string",
            "description": "The userPassword must start with a uppercase letter and contain only letters, digits, and underscores (_)"
          },
          "nickName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 8
          }
        },
        "required": [
          "userAccount",
          "userPassword"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "nickName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 8
          },
          "sex": {
            "type": "number",
            "enum": [
              0,
              1,
              2
            ],
            "description": "\n            {\n                Unknown = 0,\n                Man = 1,\n                Woman = 2\n            }\n        ",
            "default": 0
          },
          "age": {
            "type": "number"
          },
          "avatar": {
            "type": "string"
          },
          "signature": {
            "type": "string",
            "minLength": 16,
            "maxLength": 99
          },
          "hobbies": {
            "description": "['a', 'b']",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phone": {
            "type": "string",
            "description": "phone number"
          },
          "email": {
            "type": "string",
            "description": "email"
          },
          "qq": {
            "type": "string",
            "description": "qq Code"
          },
          "wechat": {
            "type": "string",
            "description": "wechat Code"
          },
          "wechatQrcode": {
            "type": "string",
            "description": "An online wechat qrcode image url is available"
          },
          "roles": {
            "description": "role id list",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "nickName",
          "sex",
          "age",
          "avatar",
          "signature",
          "hobbies",
          "email",
          "qq",
          "wechat",
          "wechatQrcode",
          "roles"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "roleName": {
            "type": "string",
            "description": "role name",
            "minLength": 2,
            "maxLength": 12
          },
          "roleKey": {
            "type": "string",
            "description": "role key",
            "minLength": 4,
            "maxLength": 12
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "maxLength": 99
          },
          "auths": {
            "description": "auth list",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roleName",
          "roleKey",
          "remark",
          "auths"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "roleName": {
            "type": "string",
            "description": "role name",
            "minLength": 2,
            "maxLength": 6
          },
          "roleKey": {
            "type": "string",
            "description": "role key",
            "minLength": 4,
            "maxLength": 8
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "maxLength": 99
          },
          "auths": {
            "description": "auth list",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roleName",
          "roleKey",
          "remark",
          "auths"
        ]
      },
...
      "CreatePageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "page title",
            "minLength": 2,
            "maxLength": 20
          },
          "path": {
            "type": "string",
            "description": "page path",
            "minLength": 4,
            "maxLength": 45
          },
          "sort": {
            "type": "number",
            "description": "page sort order",
            "minimum": 0,
            "maximum": 999
          },
          "component": {
            "type": "string",
            "description": "component path",
            "minLength": 4,
            "maxLength": 45
          },
          "isIframe": {
            "type": "string",
            "description": "is it iframe",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "type": {
            "type": "string",
            "description": "page type",
            "default": "P",
            "enum": [
              "P",
              "B"
            ]
          },
          "isLink": {
            "type": "string",
            "description": "is it external link",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "visible": {
            "type": "string",
            "description": "visible",
            "default": "0",
            "enum": [
              "0",
              "1"
            ]
          },
          "redirect": {
            "type": "string",
            "description": "redirect path",
            "minLength": 4,
            "maxLength": 45
          },
          "prefixIcon": {
            "type": "string",
            "description": "prefix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "suffixIcon": {
            "type": "string",
            "description": "suffix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "minLength": 4,
            "maxLength": 50
          },
          "auth": {
            "type": "string",
            "description": "auth id"
          }
        },
        "required": [
          "title",
          "path",
          "sort",
          "component",
          "isIframe",
          "type",
          "isLink",
          "visible",
          "redirect",
          "prefixIcon",
          "suffixIcon",
          "remark",
          "auth"
        ]
      },
      "UpdatePageDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "page title",
            "minLength": 2,
            "maxLength": 20
          },
          "path": {
            "type": "string",
            "description": "page path",
            "minLength": 4,
            "maxLength": 45
          },
          "sort": {
            "type": "number",
            "description": "page sort order",
            "minimum": 0,
            "maximum": 999
          },
          "component": {
            "type": "string",
            "description": "component path",
            "minLength": 4,
            "maxLength": 45
          },
          "isIframe": {
            "type": "string",
            "description": "is it iframe",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "type": {
            "type": "string",
            "description": "page type",
            "default": "P",
            "enum": [
              "P",
              "B"
            ]
          },
          "isLink": {
            "type": "string",
            "description": "is it external link",
            "default": "1",
            "enum": [
              "0",
              "1"
            ]
          },
          "visible": {
            "type": "string",
            "description": "visible",
            "default": "0",
            "enum": [
              "0",
              "1"
            ]
          },
          "redirect": {
            "type": "string",
            "description": "redirect path",
            "minLength": 4,
            "maxLength": 45
          },
          "prefixIcon": {
            "type": "string",
            "description": "prefix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "suffixIcon": {
            "type": "string",
            "description": "suffix icon",
            "minLength": 4,
            "maxLength": 20
          },
          "remark": {
            "type": "string",
            "description": "remark",
            "minLength": 4,
            "maxLength": 50
          },
          "auth": {
            "type": "string",
            "description": "auth id"
          }
        },
        "required": [
          "title",
          "path",
          "sort",
          "component",
          "isIframe",
          "type",
          "isLink",
          "visible",
          "redirect",
          "prefixIcon",
          "suffixIcon",
          "remark",
          "auth"
        ]
      },
...
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "tagName": {
            "type": "string",
            "description": "标签名称"
          },
          "remark": {
            "type": "string",
            "description": "标签备注"
          },
          "iconName": {
            "type": "string",
            "description": "小图标"
          },
          "postIds": {
            "description": "Post ids",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "tagName",
          "remark",
          "iconName"
        ]
      },
      "UpdateTagDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}